로컬 푸시 삭제...............................
......................................................................................

+++ Assets/0. Platform/B. Scripts/b. ScriptSystem/DataClass/AchievementData.cs
@@ -54,7 +54,7 @@ namespace PIERStory
 
             achievementSummary = SystemManager.GetJsonNodeString(__j, "summary");
 
-            achievementDegree = (float)currentPoint / (float)achievementPoint;
+            achievementDegree = (float)currentPoint / achievementPoint;
         }
     }
 }
